BugFix: show calendar on plan view when rebuild screen Refactor: split PlanTimeDataSource class to file
